Group related shopping cart methods into traits.
